Cleaned up warnings

+++ src/main/java/de/catstorm/trilife/item/DarkOrbItem.java
@@ -6,7 +6,6 @@ import net.minecraft.entity.projectile.ProjectileEntity;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ProjectileItem;
-import net.minecraft.item.WindChargeItem;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.stat.Stats;
@@ -41,9 +40,9 @@ public class DarkOrbItem extends Item implements ProjectileItem {
     @Override
     public ProjectileEntity createEntity(World world, Position pos, ItemStack stack, Direction direction) {
         Random random = world.getRandom();
-        double d = random.nextTriangular((double)direction.getOffsetX(), 0.11485000000000001);
-        double e = random.nextTriangular((double)direction.getOffsetY(), 0.11485000000000001);
-        double f = random.nextTriangular((double)direction.getOffsetZ(), 0.11485000000000001);
+        double d = random.nextTriangular(direction.getOffsetX(), 0.11485000000000001);
+        double e = random.nextTriangular(direction.getOffsetY(), 0.11485000000000001);
+        double f = random.nextTriangular(direction.getOffsetZ(), 0.11485000000000001);
         Vec3d vec3d = new Vec3d(d, e, f);
         DarkOrbEntity darkOrbEntity = new DarkOrbEntity(world, pos.getX(), pos.getY(), pos.getZ(), vec3d);
         darkOrbEntity.setVelocity(vec3d);
